pull master and edit navs
