+++ frontend/src/components/Faculty/sideNav.js
@@ -13,15 +13,20 @@ export default class SideNav extends Component {
                     <a className={`bg-green-700 capitalize p-2  mt-12 text-white  ${page === 1 && "bg-green-900"}`}
                         onClick={() => {
                             this.props.openPage(1)
-                        }}>Referrals</a>
+                        }}>Mentors</a>
                     <a className={`bg-green-700 capitalize p-2  mt-12 text-white  ${page === 2 && "bg-green-900"}`}
                         onClick={() => {
                             this.props.openPage(2)
-                        }}>Mentors</a>
-                    <a className={`bg-green-700 capitalize p-2  mt-12 text-white  ${page === 3 && "bg-green-900"}`}
-                        onClick={() => {
-                            this.props.openPage(3)
                         }}>Groups</a>
+                    <a className={`bg-green-700 capitalize p-2  mt-12 text-white  ${page === 3 && "bg-green-900"}`}
+                       onClick={() => {
+                           this.props.openPage(3)
+                       }}>Stats</a>
+                    <a className={`bg-green-700 capitalize p-2  mt-12 text-white  ${page === 4 && "bg-green-900"}`}
+                        onClick={() => {
+                            this.props.openPage(4)
+                        }}>Referrals</a>
+
                 </div>
             </div>
         )
