+++ frontend/src/components/Faculty/stats/LineChart.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {Bar, Line, Pie} from 'react-chartjs-2';
+import { Line} from 'react-chartjs-2';
 
 class LineChart extends Component{
   constructor(props){
@@ -20,8 +20,6 @@ class LineChart extends Component{
     return (
       <div className="chart">
         <Line
-          width={800}
-          height={400}
           data={this.state.chartData}
           options={{
             title:{
